1、修改订单表结构 2、订单统计

+++ litchi-module-erp/src/main/java/com/lz/module/erp/controller/admin/order/vo/OrderStatisticsRespVO.java
@@ -24,7 +24,7 @@ public class OrderStatisticsRespVO {
      * 数量
      */
     @Schema(description = "总数", requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer total;
+    private Float total;
 
 
 }
